requirements.txt
fastapi
streamlit
python-multipart
azure-ai-formrecognizer
pandas
uvicorn
requests
openpyxl
